+++ hutool-extra/src/main/java/cn/hutool/extra/pinyin/engine/tinypinyin/package-info.java
@@ -1,0 +1,22 @@
+/**
+ * 封装了TinyPinyin的引擎。
+ *
+ * <p>
+ * TinyPinyin(https://github.com/promeG/TinyPinyin)提供者未提交Maven中央库，<br>
+ * 因此使用
+ * https://github.com/biezhi/TinyPinyin打包的版本
+ * </p>
+ *
+ * <p>
+ * 引入：
+ * <pre>
+ * &lt;dependency&gt;
+ *     &lt;groupId&gt;io.github.biezhi&lt;/groupId&gt;
+ *     &lt;artifactId&gt;TinyPinyin&lt;/artifactId&gt;
+ *     &lt;version&gt;2.0.3.RELEASE&lt;/version&gt;
+ * &lt;/dependency&gt;
+ * </pre>
+ *
+ * @author looly
+ */
+package cn.hutool.extra.pinyin.engine.tinypinyin;