+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/TemplateFactory.java
@@ -1,5 +1,6 @@
 package cn.hutool.extra.template.engine;
 
+import cn.hutool.core.lang.Singleton;
 import cn.hutool.core.util.ReflectUtil;
 import cn.hutool.core.util.ServiceLoaderUtil;
 import cn.hutool.core.util.StrUtil;
@@ -14,6 +15,28 @@ import cn.hutool.log.StaticLog;
  * @author looly
  */
 public class TemplateFactory {
+
+	/**
+	 * 根据用户引入的模板引擎jar，自动创建对应的模板引擎对象<br>
+	 * 获得的是单例的TemplateEngine
+	 *
+	 * @return 单例的TemplateEngine
+	 */
+	public static TemplateEngine get(){
+		return Singleton.get(TemplateEngine.class.getName(), TemplateFactory::create);
+	}
+
+	/**
+	 * 根据用户引入的模板引擎jar，自动创建对应的模板引擎对象<br>
+	 * 推荐创建的引擎单例使用，此方法每次调用会返回新的引擎
+	 *
+	 * @return {@link TemplateEngine}
+	 * @since 5.3.3
+	 */
+	public static TemplateEngine create() {
+		return create(new TemplateConfig());
+	}
+
 	/**
 	 * 根据用户引入的模板引擎jar，自动创建对应的模板引擎对象<br>
 	 * 推荐创建的引擎单例使用，此方法每次调用会返回新的引擎